// client-monorepo/libs/applets/payment/offline-payment/src/lib/data-access/models/old-offline-payment-response.model.ts
import { ApiResultInterface } from '@client-monorepo/common/network';

export interface OldOfflinePaymentResponse {
  result: ApiResultInterface;
  trackingCode: string;
  title: string;
  imageId: string;
  titleColor: string;
  amount: number;
  providerId: string;
  time: string;
  date: string;
  creditor: OfflinePaymentCreditorResponseModel;
  creationDate: number;
  details: {
    key: string;
    value: string;
  }[];
  detailsUrl: string;
}

interface OfflinePaymentCreditorResponseModel {
  userId: string;
  name: string;
  type: number;
}